+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## To be released, v0.9.1
 
 * Add Haskell support (thanks to @nganhkhoa)
+* Include bookmarklet.js to generate testcases (Experimental)
 
 ## 2019-12-20, v0.9.0